# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Argus Personal AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
  /** @nullable */
  systemPrompt?: string | null;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  priority?: CreateTaskBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  completed?: boolean;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
  dueToday: number;
  overdue: number;
  highPriority: number;
}

export interface ResearchNote {
  id: number;
  query: string;
  summary: string;
  insights: string;
  /** @nullable */
  conclusion?: string | null;
  createdAt: string;
}

export interface CreateResearchNoteBody {
  query: string;
  summary: string;
  insights: string;
  /** @nullable */
  conclusion?: string | null;
}

export interface ResearchQueryBody {
  query: string;
}

export interface ResearchResult {
  query: string;
  summary: string;
  insights: string[];
  conclusion: string;
}

export type EmailDraftStatus =
  (typeof EmailDraftStatus)[keyof typeof EmailDraftStatus];

export const EmailDraftStatus = {
  draft: "draft",
  sent: "sent",
} as const;

export interface EmailDraft {
  id: number;
  subject: string;
  body: string;
  /** @nullable */
  recipient?: string | null;
  /** @nullable */
  context?: string | null;
  status: EmailDraftStatus;
  createdAt: string;
}

export interface CreateEmailDraftBody {
  subject: string;
  body: string;
  /** @nullable */
  recipient?: string | null;
  /** @nullable */
  context?: string | null;
}

export type GenerateEmailDraftBodyTone =
  (typeof GenerateEmailDraftBodyTone)[keyof typeof GenerateEmailDraftBodyTone];

export const GenerateEmailDraftBodyTone = {
  professional: "professional",
  casual: "casual",
  formal: "formal",
} as const;

export interface GenerateEmailDraftBody {
  context: string;
  /** @nullable */
  recipient?: string | null;
  tone?: GenerateEmailDraftBodyTone;
}

export interface GeneratedEmail {
  subject: string;
  body: string;
}

export interface SummarizeEmailBody {
  content: string;
  /** @nullable */
  sender?: string | null;
  /** @nullable */
  subject?: string | null;
}

export interface EmailSummary {
  summary: string;
  keyPoints: string[];
  actionRequired: boolean;
}

export type SocialPostTone =
  (typeof SocialPostTone)[keyof typeof SocialPostTone];

export const SocialPostTone = {
  professional: "professional",
  casual: "casual",
  promotional: "promotional",
} as const;

export interface SocialPost {
  id: number;
  topic: string;
  content: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  hashtags?: string | null;
  tone: SocialPostTone;
  /** @nullable */
  platform?: string | null;
  createdAt: string;
}

export type CreatePostBodyTone =
  (typeof CreatePostBodyTone)[keyof typeof CreatePostBodyTone];

export const CreatePostBodyTone = {
  professional: "professional",
  casual: "casual",
  promotional: "promotional",
} as const;

export interface CreatePostBody {
  topic: string;
  content: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  hashtags?: string | null;
  tone: CreatePostBodyTone;
  /** @nullable */
  platform?: string | null;
}

export type GeneratePostBodyTone =
  (typeof GeneratePostBodyTone)[keyof typeof GeneratePostBodyTone];

export const GeneratePostBodyTone = {
  professional: "professional",
  casual: "casual",
  promotional: "promotional",
} as const;

export type GeneratePostBodyLength =
  (typeof GeneratePostBodyLength)[keyof typeof GeneratePostBodyLength];

export const GeneratePostBodyLength = {
  short: "short",
  long: "long",
} as const;

export interface GeneratePostBody {
  topic: string;
  tone: GeneratePostBodyTone;
  /** @nullable */
  platform?: string | null;
  length?: GeneratePostBodyLength;
}

export interface GeneratedPost {
  content: string;
  caption: string;
  hashtags: string[];
  hook: string;
}

export interface Workflow {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  trigger: string;
  steps: string;
  enabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastRunStatus?: string | null;
  createdAt: string;
}

export interface CreateWorkflowBody {
  name: string;
  /** @nullable */
  description?: string | null;
  trigger: string;
  steps: string;
}

export interface UpdateWorkflowBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  trigger?: string;
  steps?: string;
  enabled?: boolean;
}

export interface WorkflowRunResult {
  success: boolean;
  output: string;
  stepsExecuted: number;
  /** @nullable */
  error?: string | null;
}

export interface DashboardStats {
  tasks: TaskStats;
  totalConversations: number;
  totalResearchNotes: number;
  totalPosts: number;
  totalWorkflows: number;
  activeWorkflows: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  task: "task",
  research: "research",
  email: "email",
  post: "post",
  workflow: "workflow",
  chat: "chat",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  createdAt: string;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  filter?: ListTasksFilter;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  pending: "pending",
  completed: "completed",
  all: "all",
} as const;

export type ListTasksFilter =
  (typeof ListTasksFilter)[keyof typeof ListTasksFilter];

export const ListTasksFilter = {
  today: "today",
  upcoming: "upcoming",
  all: "all",
} as const;