# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Argus Personal AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openai/conversations/{id}/messages:
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of content chunks

  # ── Tasks ──────────────────────────────────────────────────────
...
  /research:
    get:
      operationId: listResearchNotes
      tags: [research]
      summary: List research notes
      responses:
        "200":
          description: List of research notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchNote"
    post:
      operationId: createResearchNote
      tags: [research]
      summary: Save a research note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchNoteBody"
      responses:
        "201":
          description: Research note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchNote"
  /research/query:
    post:
      operationId: runResearchQuery
      tags: [research]
      summary: Run an AI research query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchQueryBody"
      responses:
        "200":
          description: Research result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResult"
  /research/{id}:
    get:
      operationId: getResearchNote
      tags: [research]
      summary: Get a research note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchNote"
        "404":
          description: Not found
    delete:
      operationId: deleteResearchNote
      tags: [research]
      summary: Delete a research note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Email Drafts ───────────────────────────────────────────────
  /emails:
    get:
      operationId: listEmailDrafts
      tags: [emails]
      summary: List email drafts
      responses:
        "200":
          description: List of email drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailDraft"
    post:
      operationId: createEmailDraft
      tags: [emails]
      summary: Save an email draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailDraftBody"
      responses:
        "201":
          description: Email draft saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /emails/draft:
    post:
      operationId: generateEmailDraft
      tags: [emails]
      summary: AI-generate an email draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailDraftBody"
      responses:
        "200":
          description: Generated draft content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"
  /emails/summarize:
    post:
      operationId: summarizeEmail
      tags: [emails]
      summary: AI-summarize email content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummarizeEmailBody"
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSummary"
  /emails/{id}:
    delete:
      operationId: deleteEmailDraft
      tags: [emails]
      summary: Delete an email draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Social Posts ───────────────────────────────────────────────
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List saved social posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: createPost
      tags: [posts]
      summary: Save a social post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
  /posts/generate:
    post:
      operationId: generatePost
      tags: [posts]
      summary: AI-generate a social media post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePostBody"
      responses:
        "200":
          description: Generated post content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPost"
  /posts/{id}:
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workflows ──────────────────────────────────────────────────
...
components:
  schemas:
...
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        systemPrompt:
          type: ["string", "null"]
      required: [content]

    # ── Tasks ──────────────────────────────────────────────────
...
    ResearchNote:
      type: object
      properties:
        id:
          type: integer
        query:
          type: string
        summary:
          type: string
        insights:
          type: string
        conclusion:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, query, summary, insights, createdAt]
    CreateResearchNoteBody:
      type: object
      properties:
        query:
          type: string
        summary:
          type: string
        insights:
          type: string
        conclusion:
          type: ["string", "null"]
      required: [query, summary, insights]
    ResearchQueryBody:
      type: object
      properties:
        query:
          type: string
      required: [query]
    ResearchResult:
      type: object
      properties:
        query:
          type: string
        summary:
          type: string
        insights:
          type: array
          items:
            type: string
        conclusion:
          type: string
      required: [query, summary, insights, conclusion]

    # ── Email Drafts ───────────────────────────────────────────
    EmailDraft:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        body:
          type: string
        recipient:
          type: ["string", "null"]
        context:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent]
        createdAt:
          type: string
      required: [id, subject, body, status, createdAt]
    CreateEmailDraftBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        recipient:
          type: ["string", "null"]
        context:
          type: ["string", "null"]
      required: [subject, body]
    GenerateEmailDraftBody:
      type: object
      properties:
        context:
          type: string
        recipient:
          type: ["string", "null"]
        tone:
          type: string
          enum: [professional, casual, formal]
      required: [context]
    GeneratedEmail:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
      required: [subject, body]
    SummarizeEmailBody:
      type: object
      properties:
        content:
          type: string
        sender:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
      required: [content]
    EmailSummary:
      type: object
      properties:
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        actionRequired:
          type: boolean
      required: [summary, keyPoints, actionRequired]

    # ── Social Posts ───────────────────────────────────────────
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        content:
          type: string
        caption:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        tone:
          type: string
          enum: [professional, casual, promotional]
        platform:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, topic, content, tone, createdAt]
    CreatePostBody:
      type: object
      properties:
        topic:
          type: string
        content:
          type: string
        caption:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        tone:
          type: string
          enum: [professional, casual, promotional]
        platform:
          type: ["string", "null"]
      required: [topic, content, tone]
    GeneratePostBody:
      type: object
      properties:
        topic:
          type: string
        tone:
          type: string
          enum: [professional, casual, promotional]
        platform:
          type: ["string", "null"]
        length:
          type: string
          enum: [short, long]
      required: [topic, tone]
    GeneratedPost:
      type: object
      properties:
        content:
          type: string
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
        hook:
          type: string
      required: [content, caption, hashtags, hook]

    # ── Workflows ──────────────────────────────────────────────
...