# lib/api-zod/src/generated/types/generatePostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Argus Personal AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratePostBodyLength } from "./generatePostBodyLength";
import type { GeneratePostBodyTone } from "./generatePostBodyTone";

export interface GeneratePostBody {
  topic: string;
  tone: GeneratePostBodyTone;
  /** @nullable */
  platform?: string | null;
  length?: GeneratePostBodyLength;
}